+++ packages/data/js/variables/color.js
@@ -4,7 +4,7 @@
  */
 import { default as memoize } from 'fast-memoize';
 import { select } from '@wordpress/data';
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
@@ -15,6 +15,7 @@ import type { ValueAddon } from '@blockera/controls/js/value-addons/types';
 /**
  * Internal dependencies
  */
+import { STORE_NAME } from '../store';
 import { generateVariableString, getBlockEditorSettings } from './index';
 import type { VariableItem } from './types';
 
@@ -43,7 +44,7 @@ export const getColors: () => Array<VariableItem> = memoize(
 				return getBlockEditorSettings()?.__experimentalFeatures?.color?.palette?.theme.map(
 					(item) => {
 						return {
-							name: item.name,
+							name: item?.name || item.slug,
 							id: item.slug,
 							value: item.color,
 							reference,
@@ -51,6 +52,24 @@ export const getColors: () => Array<VariableItem> = memoize(
 					}
 				);
 			}
+		} else if (!isUndefined(getBlockEditorSettings()?.colors)) {
+			const { getCurrentTheme } = select('blockera/data');
+
+			const theme = getCurrentTheme();
+
+			reference = {
+				type: 'theme',
+				theme: theme?.name?.rendered || '',
+			};
+
+			return getBlockEditorSettings()?.colors.map((item) => {
+				return {
+					name: item?.name || item.slug,
+					id: item.slug,
+					value: item.color,
+					reference,
+				};
+			});
 		}
 
 		if (
@@ -65,7 +84,7 @@ export const getColors: () => Array<VariableItem> = memoize(
 		return getBlockEditorSettings()?.__experimentalFeatures?.color?.palette?.default.map(
 			(item) => {
 				return {
-					name: item.name,
+					name: item?.name || item.slug,
 					id: item.slug,
 					value: item.color,
 					reference,
@@ -83,7 +102,29 @@ export const getColorsTitle: () => string = memoize(function (): string {
 					?.theme
 			)
 		) {
-			return __('Theme Colors', 'blockera');
+			const { getCurrentTheme } = select('blockera/data');
+
+			const {
+				name: { rendered: themeName },
+			} = getCurrentTheme();
+
+			return sprintf(
+				// translators: it's the product name (a theme or plugin name)
+				__('%s Colors', 'blockera'),
+				themeName
+			);
+		}
+	} else if (!isUndefined(getBlockEditorSettings()?.colors)) {
+		const { getCurrentTheme } = select('blockera/data');
+
+		const theme = getCurrentTheme();
+
+		if (!isUndefined(theme?.name?.rendered)) {
+			return sprintf(
+				// translators: it's the product name (a theme or plugin name)
+				__('%s Color Palette', 'blockera'),
+				theme?.name?.rendered
+			);
 		}
 	}
 
@@ -93,7 +134,19 @@ export const getColorsTitle: () => string = memoize(function (): string {
 export const getColor: (id: string) => ?VariableItem = memoize(function (
 	id: string
 ): ?VariableItem {
-	return getColors().find((item) => item.id === id);
+	// First, check if the color is in the default colors of theme or editor
+	let color = getColors().find((item) => item.id === id);
+
+	// If not, check if the color is in the custom colors
+	if (isUndefined(color?.value)) {
+		const { getVariableGroupItems } = select(STORE_NAME);
+
+		color = getVariableGroupItems('', 'color').find(
+			(item) => item.id === id
+		);
+	}
+
+	return color;
 });
 
 export const getColorBy: (field: string, value: any) => ?VariableItem = memoize(