+++ packages/data/js/variables/spacing.js
@@ -5,7 +5,7 @@
  */
 import { default as memoize } from 'fast-memoize';
 import { select } from '@wordpress/data';
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
@@ -44,7 +44,7 @@ export const getSpacings: () => Array<VariableItem> = memoize(
 				return getBlockEditorSettings()?.__experimentalFeatures?.spacing?.spacingSizes?.theme.map(
 					(item) => {
 						return {
-							name: item.name,
+							name: item?.name || item.slug,
 							id: item.slug,
 							value: item.size,
 							reference,
@@ -64,7 +64,7 @@ export const getSpacings: () => Array<VariableItem> = memoize(
 
 		return spaces.map((item) => {
 			return {
-				name: item.name,
+				name: item?.name || item.slug,
 				id: item.slug,
 				value: item.size,
 			};
@@ -80,7 +80,17 @@ export const getSpacingsTitle: () => string = memoize(function (): string {
 					?.spacingSizes?.theme
 			)
 		) {
-			return __('Theme Spacing Sizes', 'blockera');
+			const { getCurrentTheme } = select('blockera/data');
+
+			const theme = getCurrentTheme();
+
+			if (!isUndefined(theme?.name?.rendered)) {
+				return sprintf(
+					// translators: it's the product name (a theme or plugin name)
+					__('%s Spacing Sizes', 'blockera'),
+					theme?.name?.rendered
+				);
+			}
 		}
 	}